+++ tests/dates/test_DateDifferenceTransformer.py
@@ -19,6 +19,9 @@ from tests.dates.test_BaseGenericDateTransformer import (
     GenericDatesMixinTransformTests,
 )
 from tests.utils import (
+    _check_if_skip_test,
+    _collect_frame,
+    _convert_to_lazy,
     _handle_from_json,
     assert_frame_equal_dispatch,
     dataframe_init_dispatch,
@@ -351,6 +354,10 @@ class TestTransform(
 
     @pytest.mark.parametrize("from_json", [True, False])
     @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
+    @pytest.mark.parametrize(
         "unit",
         [
             "D",
@@ -371,7 +378,14 @@ class TestTransform(
             "polars",
         ],
     )
-    def test_expected_output_units(self, generic_expected_df, unit, library, from_json):
+    def test_expected_output_units(
+        self,
+        generic_expected_df,
+        unit,
+        library,
+        from_json,
+        lazy,
+    ):
         """Test that the output is as expected from transform, when units are D, h, m, s, week, fortnight, lunar_month, common_year, or custom_days.
 
         This tests positive month gaps, negative month gaps, and missing values.
@@ -389,10 +403,22 @@ class TestTransform(
 
         x = _handle_from_json(x, from_json)
 
-        df_transformed = x.transform(df)
+        polars = isinstance(df, pl.DataFrame)
 
-        assert_frame_equal_dispatch(expected, df_transformed)
+        if _check_if_skip_test(x, df, lazy):
+            return
 
+        df_transformed = x.transform(_convert_to_lazy(df, lazy))
+
+        assert_frame_equal_dispatch(
+            expected,
+            _collect_frame(df_transformed, polars, lazy),
+        )
+
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize("from_json", [True, False])
     @pytest.mark.parametrize(
         ("df", "expected"),
@@ -407,7 +433,7 @@ class TestTransform(
             ),
         ],
     )
-    def test_expected_output_nulls(self, df, expected, from_json):
+    def test_expected_output_nulls(self, df, expected, from_json, lazy):
         """Test that the output is expected from transform, when columns have nulls."""
         x = DateDifferenceTransformer(
             columns=["a", "b"],
@@ -416,12 +442,24 @@ class TestTransform(
             verbose=False,
         )
 
+        polars = isinstance(df, pl.DataFrame)
+
         x = _handle_from_json(x, from_json)
 
-        df_transformed = x.transform(df)
+        if _check_if_skip_test(x, df, lazy):
+            return
 
-        assert_frame_equal_dispatch(df_transformed, expected)
+        df_transformed = x.transform(_convert_to_lazy(df, lazy))
 
+        assert_frame_equal_dispatch(
+            expected,
+            _collect_frame(df_transformed, polars, lazy),
+        )
+
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize("from_json", [True, False])
     @pytest.mark.parametrize(
         ("df", "expected"),
@@ -436,7 +474,7 @@ class TestTransform(
             ),
         ],
     )
-    def test_expected_output_nulls2(self, df, expected, from_json):
+    def test_expected_output_nulls2(self, df, expected, from_json, lazy):
         """Test that the output is expected from transform, when columns are nulls."""
         x = DateDifferenceTransformer(
             columns=["a", "b"],
@@ -445,11 +483,18 @@ class TestTransform(
             verbose=False,
         )
 
+        polars = isinstance(df, pl.DataFrame)
         x = _handle_from_json(x, from_json)
 
-        df_transformed = x.transform(df)
+        if _check_if_skip_test(x, df, lazy):
+            return
 
-        assert_frame_equal_dispatch(df_transformed, expected)
+        df_transformed = x.transform(_convert_to_lazy(df, lazy))
+
+        assert_frame_equal_dispatch(
+            expected,
+            _collect_frame(df_transformed, polars, lazy),
+        )
 
 
 class TestOtherBaseBehaviour(OtherBaseBehaviourTests):