+++ tests/imputers/test_ArbitraryImputer.py
@@ -60,6 +60,10 @@ class TestTransform(
     def setup_class(cls):
         cls.transformer_name = "ArbitraryImputer"
 
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize("from_json", [True, False])
     @pytest.mark.parametrize("library", ["pandas", "polars"])
     @pytest.mark.parametrize(
@@ -77,6 +81,7 @@ class TestTransform(
         col_type,
         impute_value,
         library,
+        lazy,
         from_json,
     ):
         """Test that dtypes are preserved after imputation."""
@@ -93,6 +98,8 @@ class TestTransform(
 
         transformer = ArbitraryImputer(impute_value=impute_value, columns=[column])
 
+        if u._check_if_skip_test(transformer, df, lazy):
+            return
         transformer = _handle_from_json(transformer, from_json)
 
         if col_type in ["Categorical", "String"]:
@@ -118,8 +125,12 @@ class TestTransform(
             TypeError,
             match=msg,
         ):
-            transformer.transform(df)
+            transformer.transform(u._convert_to_lazy(df, lazy))
 
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize("from_json", [True, False])
     @pytest.mark.parametrize("library", ["pandas", "polars"])
     @pytest.mark.parametrize(
@@ -137,6 +148,7 @@ class TestTransform(
         impute_value,
         expected_values,
         library,
+        lazy,
         from_json,
     ):
         """Test that dtypes are preserved after imputation."""
@@ -155,9 +167,18 @@ class TestTransform(
 
         transformer = _handle_from_json(transformer, from_json)
 
-        df_transformed_native = transformer.transform(df_nw.to_native())
+        polars = isinstance(df, pl.DataFrame)
 
-        df_transformed_nw = nw.from_native(df_transformed_native)
+        if u._check_if_skip_test(transformer, df, lazy):
+            return
+
+        df_transformed_native = transformer.transform(
+            u._convert_to_lazy(df_nw.to_native(), lazy),
+        )
+
+        df_transformed_nw = nw.from_native(
+            u._collect_frame(df_transformed_native, polars, lazy),
+        )
 
         expected_dtype = df_nw[column].dtype
 
@@ -184,13 +205,17 @@ class TestTransform(
 
         u.assert_frame_equal_dispatch(
             expected.to_native(),
-            df_transformed_native,
+            df_transformed_nw.to_native(),
             # this turns off checks for category metadata like ordering
             # this transformer will convert an unordered pd categorical to ordered
             # so this is needed
             check_categorical=False,
         )
 
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize("from_json", [True, False])
     @pytest.mark.parametrize("library", ["pandas", "polars"])
     @pytest.mark.parametrize(
@@ -207,6 +232,7 @@ class TestTransform(
         impute_value,
         expected_values,
         library,
+        lazy,
         from_json,
     ):
         """Test handling for some edge cases:
@@ -223,6 +249,10 @@ class TestTransform(
 
         transformer = ArbitraryImputer(impute_value=impute_value, columns=[column])
 
+        polars = isinstance(df, pl.DataFrame)
+
+        if u._check_if_skip_test(transformer, df, lazy):
+            return
         transformer = _handle_from_json(transformer, from_json)
 
         # for pandas, the all null column is inferred as string type
@@ -232,12 +262,18 @@ class TestTransform(
                 UserWarning,
                 match=f"{self.transformer_name}: X contains all null columns { {column}!s}, types for these columns will be inferred as {type(transformer.impute_value)}",
             ):
-                df_transformed_native = transformer.transform(df_nw.to_native())
+                df_transformed_native = transformer.transform(
+                    u._convert_to_lazy(df_nw.to_native(), lazy),
+                )
 
         else:
-            df_transformed_native = transformer.transform(df_nw.to_native())
+            df_transformed_native = transformer.transform(
+                u._convert_to_lazy(df_nw.to_native(), lazy),
+            )
 
-        df_transformed_nw = nw.from_native(df_transformed_native)
+        df_transformed_nw = nw.from_native(
+            u._collect_frame(df_transformed_native, polars, lazy),
+        )
 
         actual_dtype = str(df_transformed_nw[column].dtype)
 
@@ -253,9 +289,16 @@ class TestTransform(
             ),
         )
 
-        u.assert_frame_equal_dispatch(expected.to_native(), df_transformed_native)
+        u.assert_frame_equal_dispatch(
+            expected.to_native(),
+            df_transformed_nw.to_native(),
+        )
 
     @pytest.mark.parametrize("from_json", [True, False])
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize(
         ("impute_value", "impute_val_type"),
         [
@@ -264,7 +307,13 @@ class TestTransform(
             (True, "Boolean"),
         ],
     )
-    def test_polars_unknown_type_output(self, impute_value, impute_val_type, from_json):
+    def test_polars_unknown_type_output(
+        self,
+        impute_value,
+        impute_val_type,
+        lazy,
+        from_json,
+    ):
         """Test handling of polars Unknown type column (output type should be inferred from impute_value)"""
 
         column = "a"
@@ -277,11 +326,19 @@ class TestTransform(
 
         transformer = ArbitraryImputer(impute_value=impute_value, columns=[column])
 
+        polars = isinstance(df, pl.DataFrame)
         transformer = _handle_from_json(transformer, from_json)
 
-        df_transformed_native = transformer.transform(df_nw.to_native())
+        if u._check_if_skip_test(transformer, df, lazy):
+            return
 
-        df_transformed_nw = nw.from_native(df_transformed_native)
+        df_transformed_native = transformer.transform(
+            u._convert_to_lazy(df_nw.to_native(), lazy),
+        )
+
+        df_transformed_nw = nw.from_native(
+            u._collect_frame(df_transformed_native, polars, lazy),
+        )
 
         actual_dtype = str(df_transformed_nw[column].dtype)
 
@@ -299,10 +356,17 @@ class TestTransform(
             ).cast(getattr(nw, impute_val_type)),
         )
 
-        u.assert_frame_equal_dispatch(expected.to_native(), df_transformed_native)
+        u.assert_frame_equal_dispatch(
+            expected.to_native(),
+            df_transformed_nw.to_native(),
+        )
 
     # have to overload this one, as has slightly different categorical type handling
     @pytest.mark.parametrize("from_json", [True, False])
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize(
         ("library", "expected_df_3", "impute_values_dict"),
         [
@@ -317,6 +381,7 @@ class TestTransform(
         expected_df_3,
         initialized_transformers,
         impute_values_dict,
+        lazy,
         from_json,
     ):
         """Test that transform is giving the expected output when applied to object and categorical columns."""
@@ -325,6 +390,11 @@ class TestTransform(
 
         # Initialize the transformer
         transformer = initialized_transformers[self.transformer_name]
+
+        polars = isinstance(df2, pl.DataFrame)
+
+        if u._check_if_skip_test(transformer, df2, lazy):
+            return
 
         transformer.impute_values_ = impute_values_dict
 
@@ -336,11 +406,11 @@ class TestTransform(
         transformer = _handle_from_json(transformer, from_json)
 
         # Transform the DataFrame
-        df_transformed = transformer.transform(df2)
+        df_transformed = transformer.transform(u._convert_to_lazy(df2, lazy))
 
         # Check whole dataframes
         u.assert_frame_equal_dispatch(
-            df_transformed,
+            u._collect_frame(df_transformed, polars, lazy),
             expected_df_3,
             # this turns off checks for category metadata like ordering
             # this transformer will convert an unordered pd categorical to ordered
@@ -351,11 +421,13 @@ class TestTransform(
         expected_df_3 = nw.from_native(expected_df_3)
 
         for i in range(len(df2)):
-            df_transformed_row = transformer.transform(df2[[i]].to_native())
+            df_transformed_row = transformer.transform(
+                u._convert_to_lazy(df2[[i]].to_native(), lazy),
+            )
             df_expected_row = expected_df_3[[i]].to_native()
 
             u.assert_frame_equal_dispatch(
-                df_transformed_row,
+                u._collect_frame(df_transformed_row, polars, lazy),
                 df_expected_row,
                 # this turns off checks for category metadata like ordering
                 # this transformer will convert an unordered pd categorical to ordered
@@ -364,6 +436,10 @@ class TestTransform(
             )
 
     @pytest.mark.parametrize("from_json", [True, False])
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize(
         ("library", "expected_df_4", "impute_values_dict"),
         [
@@ -378,6 +454,7 @@ class TestTransform(
         expected_df_4,
         initialized_transformers,
         impute_values_dict,
+        lazy,
         from_json,
     ):
         """Test that transform is giving the expected output when applied to object and categorical columns
@@ -389,6 +466,11 @@ class TestTransform(
         # Initialize the transformer
         transformer = initialized_transformers[self.transformer_name]
 
+        polars = isinstance(df2, pl.DataFrame)
+
+        if u._check_if_skip_test(transformer, df2, lazy):
+            return
+
         transformer.impute_values_ = impute_values_dict
         transformer.impute_value = "z"
         transformer.columns = ["b", "c"]
@@ -396,11 +478,11 @@ class TestTransform(
         transformer = _handle_from_json(transformer, from_json)
 
         # Transform the DataFrame
-        df_transformed = transformer.transform(df2)
+        df_transformed = transformer.transform(u._convert_to_lazy(df2, lazy))
 
         # Check whole dataframes
         u.assert_frame_equal_dispatch(
-            df_transformed,
+            u._collect_frame(df_transformed, polars, lazy),
             expected_df_4,
             # this turns off checks for category metadata like ordering
             # this transformer will convert an unordered pd categorical to ordered
@@ -411,11 +493,13 @@ class TestTransform(
         expected_df_4 = nw.from_native(expected_df_4)
 
         for i in range(len(df2)):
-            df_transformed_row = transformer.transform(df2[[i]].to_native())
+            df_transformed_row = transformer.transform(
+                u._convert_to_lazy(df2[[i]].to_native(), lazy),
+            )
             df_expected_row = expected_df_4[[i]].to_native()
 
             u.assert_frame_equal_dispatch(
-                df_transformed_row,
+                u._collect_frame(df_transformed_row, polars, lazy),
                 df_expected_row,
                 # this turns off checks for category metadata like ordering
                 # this transformer will convert an unordered pd categorical to ordered
@@ -423,6 +507,10 @@ class TestTransform(
                 check_categorical=False,
             )
 
+    @pytest.mark.parametrize(
+        "lazy",
+        [True, False],
+    )
     @pytest.mark.parametrize("from_json", [True, False])
     @pytest.mark.parametrize("library", ["pandas", "polars"])
     @pytest.mark.parametrize(
@@ -436,6 +524,7 @@ class TestTransform(
         self,
         input_values,
         library,
+        lazy,
         from_json,
     ):
         """Test that unexpected dtypes will hit error"""
@@ -448,6 +537,8 @@ class TestTransform(
 
         transformer = ArbitraryImputer(impute_value=1, columns=[column])
 
+        if u._check_if_skip_test(transformer, df, lazy):
+            return
         transformer = _handle_from_json(transformer, from_json)
 
         bad_types = dict(nw.from_native(df).select(nw.col(column)).schema.items())
@@ -463,7 +554,7 @@ class TestTransform(
             TypeError,
             match=msg,
         ):
-            transformer.transform(df)
+            transformer.transform(u._convert_to_lazy(df, lazy))
 
 
 class TestOtherBaseBehaviour(OtherBaseBehaviourTests):