+++ tubular/imputers.py
@@ -67,7 +67,7 @@ class BaseImputer(BaseTransformer):
 
     polars_compatible = True
 
-    lazyframe_compatible = False
+    lazyframe_compatible = True
 
     # this class is not by itself jsonable, as needs attrs
     # which are set in the child classes
@@ -274,7 +274,7 @@ class ArbitraryImputer(BaseImputer):
 
     polars_compatible = True
 
-    lazyframe_compatible = False
+    lazyframe_compatible = True
 
     jsonable = True
 
@@ -1092,7 +1092,7 @@ class NullIndicator(BaseTransformer):
 
     polars_compatible = True
 
-    lazyframe_compatible = False
+    lazyframe_compatible = True
 
     FITS = False