+++ tubular/misc.py
@@ -49,7 +49,7 @@ class SetValueTransformer(BaseTransformer):
 
     polars_compatible = True
 
-    lazyframe_compatible = False
+    lazyframe_compatible = True
 
     FITS = False