+++ tubular/numeric.py
@@ -85,7 +85,7 @@ class BaseNumericTransformer(BaseTransformer, CheckNumericMixin):
 
     polars_compatible = True
 
-    lazyframe_compatible = False
+    lazyframe_compatible = True
 
     jsonable = False
 
@@ -127,7 +127,7 @@ class BaseNumericTransformer(BaseTransformer, CheckNumericMixin):
 
         super().fit(X, y)
 
-        CheckNumericMixin.check_numeric_columns(self, X[self.columns])
+        CheckNumericMixin.check_numeric_columns(self, X.select(self.columns))
 
         return self
 
@@ -489,7 +489,7 @@ class DifferenceTransformer(BaseNumericTransformer):
 
     jsonable = True
 
-    lazyframe_compatible = False
+    lazyframe_compatible = True
 
     @beartype
     def __init__(
@@ -613,7 +613,7 @@ class RatioTransformer(BaseNumericTransformer):
 
     jsonable = True
 
-    lazyframe_compatible = False
+    lazyframe_compatible = True
 
     @block_from_json
     def to_json(self) -> dict[str, dict[str, Any]]: